Add Triangle component with a preview
This commit adds a new file called "Triangle.swift" which contains the implementation of the Triangle component. The Triangle struct conforms to the View protocol and displays the text "Hello, World!". Additionally, a preview for the Triangle component is included.
Add Contacts button to toolbar

This commit adds a new Contacts button to the toolbar. The button is represented by an image of a person with a plus sign and is styled with green and blue colors. When tapped, it presents a popover view called "ContactsSelectionView" which allows the user to select contacts.